pyproject.toml
[tool.poetry]
name = "atsc"
version = "0.1.0"
description = "Acoustic Traffic Simulation and Counting"
authors = [
    "Luca Bondi <luca.bondi@us.bosch.com>",
    "Stefano Damiano <stefano.damiano@ekuleuven.be>",
]
maintainers = [
    "Luca Bondi <luca.bondi@us.bosch.com>",
    "Winston Lin <winston.lin@us.bosch.com>",
    "Shabnam Ghaffarzadegan <shabnam.ghaffarzadegan@us.bosch.com>",
    "Abinaya Kumar <abinaya.kumar@us.bosch.com>",
]
readme = "README.md"
license = "GPL-3.0-only"
repository = "https://github.com/boschresearch/acoustic-traffic-simulation-counting"
keywords = [
    "traffic counting",
    "acoustic simulation"
]
classifiers = [
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Intended Audience :: Education",
    "Environment :: GPU :: NVIDIA CUDA",
    "Development Status :: 4 - Beta",
]

[tool.poetry.dependencies]
python = "^3.10"
pyroadacoustics = "^1.1.0"
torch = "^2.1.2"
torchaudio = "^2.1.2"
numpy = "^1.26.3"
scipy = "^1.12.0"
hydra-core = "^1.3.2"
lightning = "^2.1.3"
librosa = "^0.10.1"
torchmetrics = "^1.3.0.post0"
matplotlib = "^3.8.2"
pandas = "^2.2.0"
python-dotenv = "^1.0.1"
neptune = {version = "^1.9.1", optional = true}
mlflow = {version = "^2.10.2", optional = true}
coolname = "^2.2.0"

[tool.poetry.extras]
mlflow = ["mlflow"]
neptune = ["neptune"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
